+++ pages/_app.js
@@ -1,31 +1,13 @@
-import '../styles/_app/destyle.css'
-import '../styles/_app/globals.scss'
-import Head from 'next/head'
+import usePageView from '../hooks/usePageView'
 import Header from '../components/Header'
 import Footer from '../components/Footer'
-import * as gtag from '~/src/lib/gtag'
+import '../styles/_app/destyle.css'
+import '../styles/_app/globals.scss'
 
 function MyApp({ Component, pageProps }) {
-  useEffect(() => {
-    if (!gtag.existsGaId) {
-      return
-    }
-
-    const handleRouteChange = (path) => {
-      gtag.pageview(path)
-    }
-
-    router.events.on('routeChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
+  usePageView()
   return (
     <>
-      <Head>
-        <link rel="preconnect" href="https://fonts.gstatic.com" />
-        <link href="https://fonts.googleapis.com/css2?family=Damion&display=swap" rel="stylesheet" />
-      </Head>
       <Header />
       <Component {...pageProps} />
       <Footer />
