+++ pages/_document.js
@@ -6,6 +6,8 @@ export default class MyDocument extends Document {
     return (
       <html lang="ja">
         <Head>
+          <link rel="preconnect" href="https://fonts.gstatic.com" />
+          <link href="https://fonts.googleapis.com/css2?family=Damion&display=swap" rel="stylesheet" />
           {/* Google Analytics */}
           {existsGaId && (
             <>
